requirements.txt
# set conda python 3.9
# conda install tensorflow-gpu #it says no need gpu tag anymore, but we still need 2.6.2 needed
# CondaSSLError: OpenSSL appears to be unavailable <- add conda to path
# conda install -c conda-forge ffmpeg
# conda install keras
# pip install -r requirements.txt to install below
tensorflow-datasets==4.4.0
tensorflow-addons==0.15.0
absl-py==0.12.0
gin-config==0.5.0
parameterized==0.8.1
mediapy==1.0.3
scikit-image==0.19.1
apache-beam==2.34.0
google-cloud-bigquery-storage==1.1.0 # Suppresses a harmless error from beam
natsort==8.1.0
gdown==4.5.4
tqdm==4.64.1
python-snappy
notebook
pandas
#numpy==1.23.4 if np.object error occurs